Fix: Order-Article-Mapping; @Table Annotationen entfernt

+++ src/main/java/com/fhdw/loeppe/repo/ArticleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {}
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {}
 
 //TODO: RepositoryTest
